Migrate main to pkg.
