+++ pkg/unifi-poller/jsondebug.go
@@ -1,4 +1,4 @@
-package main
+package unifipoller
 
 import (
 	"fmt"
